requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.65.0

# NLP and ML libraries
spacy>=3.5.0
nltk>=3.8.1
scikit-learn>=1.0.2
sentence-transformers>=2.2.2

# GUI dependencies
ttkthemes>=3.2.2
Pillow>=9.3.0

# Data handling
pyyaml>=6.0
json5>=0.9.14

# Utility libraries
pathlib>=1.0.1
python-dateutil>=2.8.2

# Optional dependencies for enhanced features
# sounddevice>=0.4.6  # For voice input feature (uncomment when implemented)
# soundfile>=0.12.1   # For voice input feature (uncomment when implemented)

# Development dependencies
black>=22.3.0  # Code formatting
pylint>=2.15.0  # Code linting
pytest>=7.3.1  # Testing

language-tool-python==2.7.0
regex==2023.10.3